+++ schemas.py
@@ -78,6 +78,16 @@ class Schema_Probe(BaseModel):
 class Schema_Just_UTI_SN(BaseModel):
     sn: int
 
+class Schema_Calibration(BaseModel):
+    sn: int
+    sens_meth: str
+    ref_size: str
+    ref_level: str
+    trans_cor: str
+    scan_lev: str
+    screen_range: str
+    scan_type: str
+
 class Schema_Report(BaseModel):
 
     test: str | None=None 
@@ -89,15 +99,14 @@ class Schema_Report(BaseModel):
     heat: str | None=None 
     j_order: str  | None=None
     j_qty: str 
-    od: float 
-    id: float 
-    width: float 
-    height: float 
+    od: str 
+    id: str 
+    width: str
+    height: str
     NDE: str 
     crit_accept: str | None=None 
     rough: str | None=None
     uti_sn: int
-    sn1: int
     d_cal: str
     sens_block: str
     notch: str
@@ -114,8 +123,80 @@ class Schema_Report(BaseModel):
     cert_due: str| None = None
     acc_sn: int | None = None
     rej_sn: int | None = None
-    sn2: int | None = None
-    sn3: int | None = None
-    sn4: int | None = None
-    sn5: int | None = None
-    
+    calibrations: list
+
+# class Schema_Report(BaseModel):
+
+#     test: str | None=None 
+#     client_name: str  | None=None
+#     plant: str | None=None 
+#     contact_name: str  | None=None
+#     part_desc: str 
+#     material: str 
+#     heat: str | None=None 
+#     j_order: str  | None=None
+#     j_qty: str 
+#     od: str 
+#     id: str 
+#     width: str
+#     height: str
+#     NDE: str 
+#     crit_accept: str | None=None 
+#     rough: str | None=None
+#     uti_sn: int
+#     sn1: int
+#     d_cal: str
+#     sens_block: str
+#     notch: str
+#     rec_lvl: str
+#     ax_scanning: str
+#     circ_ax_scanning: str
+#     method: str
+#     coupling: str
+#     stage: str
+#     remarks: str
+#     insp_name: str
+#     ndt_act: str | None=None
+#     cert_lvl: str | None = None
+#     cert_due: str| None = None
+#     acc_sn: int | None = None
+#     rej_sn: int | None = None
+#     sn2: int | None = None
+#     sn3: int | None = None
+#     sn4: int | None = None
+#     sn5: int | None = None
+#     sens_meth1: str
+#     sens_meth2: str | None = None
+#     sens_meth3: str | None = None
+#     sens_meth4: str | None = None
+#     sens_meth5: str | None = None
+#     ref_size1: str
+#     ref_size2: str | None = None
+#     ref_size3: str | None = None
+#     ref_size4: str | None = None
+#     ref_size5: str | None = None
+#     ref_level1: str
+#     ref_level2: str | None = None
+#     ref_level3: str | None = None
+#     ref_level4: str | None = None
+#     ref_level5: str | None = None
+#     trans_cor1: str
+#     trans_cor2: str | None = None
+#     trans_cor3: str | None = None
+#     trans_cor4: str | None = None
+#     trans_cor5: str | None = None
+#     scan_lev1: str
+#     scan_lev2: str | None = None
+#     scan_lev3: str | None = None
+#     scan_lev4: str | None = None
+#     scan_lev5: str | None = None
+#     screen_range1: str
+#     screen_range2: str | None = None
+#     screen_range3: str | None = None
+#     screen_range4: str | None = None
+#     screen_range5: str | None = None
+#     scan_type1: str
+#     scan_type2: str | None = None
+#     scan_type3: str | None = None
+#     scan_type4: str | None = None
+#     scan_type5: str | None = None